+++ vnc_winsrc/winvnc/vncEncodeRRE.cpp
@@ -1,3 +1,4 @@
+//  Copyright (C) 2010 D. R. Commander. All Rights Reserved.
 //  Copyright (C) 2000 Tridia Corporation. All Rights Reserved.
 //  Copyright (C) 1999 AT&T Laboratories Cambridge. All Rights Reserved.
 //
@@ -298,7 +299,7 @@ vncEncodeRRE::EncodeRect(BYTE *source, BYTE *dest, const RECT &rect, int offx, i
 			dest+sz_rfbFramebufferUpdateRectHeader+sz_rfbRREHeader,
 			rectW,
 			rectH,
-			m_bufflen-sz_rfbFramebufferUpdateRectHeader-sz_rfbRREHeader
+			(int)(m_bufflen-sz_rfbFramebufferUpdateRectHeader-sz_rfbRREHeader)
 			);
 		break;
 	case 16:
@@ -307,7 +308,7 @@ vncEncodeRRE::EncodeRect(BYTE *source, BYTE *dest, const RECT &rect, int offx, i
 			(CARD8 *)(dest+sz_rfbFramebufferUpdateRectHeader+sz_rfbRREHeader),
 			rectW,
 			rectH,
-			m_bufflen-sz_rfbFramebufferUpdateRectHeader-sz_rfbRREHeader
+			(int)(m_bufflen-sz_rfbFramebufferUpdateRectHeader-sz_rfbRREHeader)
 			);
 		break;
 	case 32:
@@ -316,7 +317,7 @@ vncEncodeRRE::EncodeRect(BYTE *source, BYTE *dest, const RECT &rect, int offx, i
 			(CARD8 *)(dest+sz_rfbFramebufferUpdateRectHeader+sz_rfbRREHeader),
 			rectW,
 			rectH,
-			m_bufflen-sz_rfbFramebufferUpdateRectHeader-sz_rfbRREHeader
+			(int)(m_bufflen-sz_rfbFramebufferUpdateRectHeader-sz_rfbRREHeader)
 			);
 		break;
 	}
